+++ NEWS.md
@@ -1,4 +1,4 @@
-# crew 0.7.0.9011 (development)
+# crew 0.7.0.9012 (development)
 
 * Configure workers to send themselves a termination signal if the connection to the dispatcher is broken (#141, @psychelzh). Huge thanks to @shikokuchuo for the support through https://github.com/shikokuchuo/mirai/issues/87, https://github.com/shikokuchuo/mirai/pull/88, and https://github.com/shikokuchuo/nanonext/pull/25! The signal itself is platform-dependent and determined by the new function `crew_terminate_signal()`.
 * Implement `crew_monitor_local()` to help users monitor and terminate local R processes created by `crew` and `mirai`.